+++ hutool-poi/src/main/java/cn/hutool/poi/excel/sax/ExcelSaxUtil.java
@@ -73,7 +73,11 @@ public class ExcelSaxUtil {
 				}
 				break;
 			case NUMBER:
-				result = getNumberValue(value, numFmtString);
+				try{
+					result = getNumberValue(value, numFmtString);
+				}catch (NumberFormatException e){
+					result = value;
+				}
 				break;
 			case DATE:
 				try {